fix: resolve all clippy warnings
- Replace manual_map patterns with Option::map
- Fix redundant closures with method references
- Change &PathBuf to &Path in function signatures
- Replace .min().max() with .clamp()
- Use .div_ceil() instead of manual division
- Replace .last() with .next_back() on DoubleEndedIterator
- Fix collapsible_if patterns
- Remove unnecessary clone() on Copy types
- Replace let-else with ? operator where applicable
- Use is_ok_and() instead of map_or() for Result checks
- Use iter().flatten() instead of manual Option handling
- Remove unnecessary casts and borrows
- Use repeat_n() instead of repeat().take()
- Remove redundant max(0) on unsigned types
- Rename from_str to parse to avoid should_implement_trait warning

+++ src/ui/components/input_box.rs
@@ -759,7 +759,7 @@ impl InputBox {
             return;
         }
 
-        let relative_x = (click_x - content_x) as u16;
+        let relative_x = click_x - content_x;
         let relative_y = (click_y - content_y) as usize;
 
         let visual_lines = if show_scrollbar {
